refactor(test): extract pagination-test-helper
